requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
greenlet==1.1.2
idna==3.3
lxml==4.9.0
numpy==1.23.0
pandas==1.4.3
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
urllib3==1.26.9
